requirements.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
Pillow==9.0.1
pycodestyle==2.8.0
requests==2.27.1
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.8
